collect logs from openshift-sdn namespace
